Fix custombuild invocation under Linux.
shebang line interpretation has different behavior across Unix flavors.
Some will break the given command line and call exec with
correct arguments, but some will just take everything between "#!" and
"\n" and pass it as the first argument to exec, leading to process
execution failure. This behavior occurs notably in Linux, which causes
custombuild executions to fail.

The only portable way is to split the script in two files: one
containing the actual script code and the other be only a shell script
invocating the former with correct command line.